+++ config/viewer.go
@@ -15,7 +15,6 @@ type ViewerConfig struct {
 	ParseHttpLinks bool       `ini:"parse-http-links"`
 	HeaderLayout   [][]string `ini:"-"`
 	KeyPassthrough bool       `ini:"-"`
-	CloseOnReply   bool       `ini:"close-on-reply"`
 }
 
 func defaultViewerConfig() *ViewerConfig {
@@ -30,7 +29,6 @@ func defaultViewerConfig() *ViewerConfig {
 			{"Subject"},
 		},
 		ParseHttpLinks: true,
-		CloseOnReply:   false,
 	}
 }